Move legacy code to legacy file to allow for backward compatibility
